requirements.txt
Django==1.10.4
anyjson==0.3.3
djangorestframework==3.5.3
psycopg2==2.6.2
pytz==2016.7
requests==2.12.3
vine==1.1.3
wsgiref==0.1.2
